requirements.txt
fastapi==0.110.2
pydantic-settings==2.2.1
pydantic==2.7.1
uvicorn==0.29.0
openai==1.23.6
pyjwt==2.8.0
requests==2.31.0
